🧪 😷 Let's take care of each other
In case of any cold symptoms during the event,
we suggest that you wear an FFP2 mask whenever possible.
In addition, you might do a self-test for Covid.
Feel free to approach the conference organisers for mask(s) and/or a test.
Of course you can also choose to wear a mask without symptoms.
We will strive for good air quality in the rooms.